% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbr_bounds}
\alias{gbr_bounds}
\title{Great Barrier Reef Marine Park Boundary}
\format{
A \href{https://cran.r-project.org/package=sf}{Simple Features data.frame}
with 1 row and 13 variables:
\itemize{
\item OBJECTID:
\item GIS_AREA:
\item sdc_extern:
\item HECTARES:
\item UNIQUE_ID:
\item LOC_TYPE_S:
\item LOC_NAME_S:
\item LOC_NAM_L:
\item SHAPE_Leng:
\item GlobalID:
\item Shape_area:
\item Shape_len:
\item geometry:
}
}
\description{
Derived from the Amalgamated Great Barrier Reef (AGBR) Section that
replaces all previously existing  sections.The AGBR was proclaimed
under subsection 31 (1) of the Act in 2003 (Gazette No S119) dated 21
April 2004.
}
\details{
For full Metadata see the original
\href{https://tinyurl.com/y93w8euw}{GBRMPA Geoportal}
page info for the Great Barrier Reef Marine Park Boundary.

Data kindly provided by:
Great Barrier Reef Marine Park Authority
Data Administrator
Spatial Data Centre
PO Box 1379
Townsville, QLD 4810
Australia
\email{gis@gbrmpa.gov.au}

The Great Barrier Reef Marine Park Authority requests attribution
in the following manner:
Copyright (c) Great Barrier Reef Marine Park Authority 2020.
This product is released under
the \href{http://creativecommons.org/licenses/by/4.0/legalcode}{Creative Commons Attribution 4.0 International Licence}.
Updated data available at \url{http://www.gbrmpa.gov.au/geoportal/}.
Under CC BY 4.0, users are free to share and adapt the
data under the following conditions:
\itemize{
\item They must give appropriate credit
\item Provide a link to the license
\item Indicate if any changes were made, and
\item They may not apply any additional restrictions that
legally restrict others from doing anything the license permits.
}
}
