% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbr_feat}
\alias{gbr_feat}
\title{Great Barrier Reef Features}
\format{
A \href{https://cran.r-project.org/package=sf}{Simple Features data.frame}
with 5,363 rows and 21 variables:
\itemize{
\item OBJECTID:
\item SORT_GBR_I:
\item LABEL_ID:
\item SUB_NO:
\item CODE:
\item UNIQUE_ID:
\item FEATURE_C:
\item GBR_NAME:
\item FEAT_NAME:
\item QLD_NAME:
\item X_LABEL:
\item GBR_ID:
\item LOC_NAME_S:
\item LOC_NAME_L:
\item X_COORD:
\item Y_COORD:
\item Area_HA:
\item GlobalID:
\item Shape_STAr:
\item Shape_STLe:
\item geometry:
}
}
\description{
Coastal features within and adjacent to the Great Barrier Reef
World Heritage area (such as islands, rocks and the Queensland
mainland). - Major coral cay features - Major and other coral reef
structures (as defined by the reef shoal edge) and tidal, drying or
emergent reef areas within the Great Barrier Reef World Heritage Area.
All data is based on the GDA2020 coordinate system.
}
\details{
For full Metadata see the original
\href{https://tinyurl.com/y9mltz6n}{GBRMPA Geoportal}
page info for the Great Barrier Reef Features.

Data kindly provided by:
Great Barrier Reef Marine Park Authority
Data Administrator
Spatial Data Centre
PO Box 1379
Townsville, QLD 4810
Australia
\email{gis@gbrmpa.gov.au}

The Great Barrier Reef Marine Park Authority requests attribution
in the following manner:
Copyright (c) Great Barrier Reef Marine Park Authority 2020.
This product is released under
the \href{http://creativecommons.org/licenses/by/4.0/legalcode}{Creative Commons Attribution 4.0 International Licence}.
Updated data available at \url{http://www.gbrmpa.gov.au/geoportal/}.
Under CC BY 4.0, users are free to share and adapt the
data under the following conditions:
\itemize{
\item They must give appropriate credit
\item Provide a link to the license
\item Indicate if any changes were made, and
\item They may not apply any additional restrictions that
legally restrict others from doing anything the license permits.
}
}
