% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrm_regions}
\alias{nrm_regions}
\title{Name of file}
\format{
A \href{https://cran.r-project.org/package=sf}{Simple Features data.frame}
with XX rows and XX variables:
\itemize{
\item 1:
\item 2:
}
}
\description{
Lines of general description
Lines of general description
Lines of general description
}
\details{
For full Metadata see the original
GBRMPA Geoportal (add link)
page info for the Data X.

Data kindly provided by:
Great Barrier Reef Marine Park Authority
Data Administrator
Spatial Data Centre
PO Box 1379
Townsville, QLD 4810
Australia
\email{gis@gbrmpa.gov.au}

The Great Barrier Reef Marine Park Authority requests attribution
in the following manner:
Copyright (c) Great Barrier Reef Marine Park Authority 2020.
This product is released under
the \href{http://creativecommons.org/licenses/by/4.0/legalcode}{Creative Commons Attribution 4.0 International Licence}.
Updated data available at \url{http://www.gbrmpa.gov.au/geoportal/}.
Under CC BY 4.0, users are free to share and adapt the
data under the following conditions:
\itemize{
\item They must give appropriate credit
\item Provide a link to the license
\item Indicate if any changes were made, and
\item They may not apply any additional restrictions that
legally restrict others from doing anything the license permits.
}
}
