% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qld_hydro}
\alias{qld_hydro}
\title{Name of file}
\format{
A \href{https://cran.r-project.org/package=sf}{Simple Features data.frame}
with XX rows and XX variables:
\itemize{
\item 1:
\item 2:
}
}
\description{
Lines of general description
Lines of general description
Lines of general description
}
\details{
For full Metadata see the original
Geoscience Australia Data Portal
page info for the \href{http://pid.geoscience.gov.au/dataset/ga/64058}{GEODATA TOPO 250K Series 3}.

Data kindly provided by:
Great Barrier Reef Marine Park Authority
Data Administrator
Spatial Data Centre
PO Box 1379
Townsville, QLD 4810
Australia
\email{gis@gbrmpa.gov.au}

Geoscience Australia requests attribution
in the following manner:
Copyright (c) Commonwealth of Australia (Geoscience Australia) 2020.
This product is released under the \href{http://creativecommons.org/licenses/by/4.0/legalcode}{Creative Commons Attribution 4.0 International Licence}

Publication details:
\itemize{
\item This product was created on and published on Jan 1, 2006
\item Persistent Identifier: http://pid.geoscience.gov.au/dataset/ga/64058
\item Citation: 2006. GEODATA TOPO 250K Series 3 (Shape file format). Geoscience Australia, Canberra. http://pid.geoscience.gov.au/dataset/ga/64058
\item Legal Constraints: \href{http://creativecommons.org/licenses/by/4.0/legalcode}{Creative Commons Attribution 4.0 International Licence}
\item Access Constraints: \href{http://creativecommons.org/licenses/by/4.0/legalcode}{Creative Commons Attribution 4.0 International Licence}
\item Use Constraints: \href{http://creativecommons.org/licenses/by/4.0/legalcode}{Creative Commons Attribution 4.0 International Licence}
}

Under CC BY 4.0, users are free to share and adapt the
data under the following conditions:
\itemize{
\item They must give appropriate credit
\item Provide a link to the license
\item Indicate if any changes were made, and
\item They may not apply any additional restrictions that
legally restrict others from doing anything the license permits.
}
}
