% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wbodies}
\alias{wbodies}
\title{Water Bodies}
\format{
A \href{https://cran.r-project.org/package=sf}{Simple Features data.frame}
with 485 rows and 16 variables:
\itemize{
\item typesource:
\item MarineWate:
\item UniqueID:
\item chl_annnua:
\item chl_50th:
\item chl_80th:
\item TSS_annual:
\item TSS_50th:
\item TSS_80th:
\item chl_source:
\item TSS_source:
\item Lat_DD:
\item Lon_DD:
\item Shape_STAr:
\item Shape_STLe:
\item geometry:
}
}
\description{
Description
}
\details{
Data kindly provided by:
Great Barrier Reef Marine Park Authority
Data Administrator
Spatial Data Centre
PO Box 1379
Townsville, QLD 4810
Australia
\email{gis@gbrmpa.gov.au}

The Great Barrier Reef Marine Park Authority requests attribution
in the following manner:
Copyright (c) Great Barrier Reef Marine Park Authority 2020.
This product is released under
the \href{http://creativecommons.org/licenses/by/4.0/legalcode}{Creative Commons Attribution 4.0 International Licence}.
Updated data available at \url{http://www.gbrmpa.gov.au/geoportal/}.
Under CC BY 4.0, users are free to share and adapt the
data under the following conditions:
\itemize{
\item They must give appropriate credit
\item Provide a link to the license
\item Indicate if any changes were made, and
\item They may not apply any additional restrictions that
legally restrict others from doing anything the license permits.
}
}
