% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrm_regions}
\alias{nrm_regions}
\title{Natural Resource Management Marine Regions Relating to the
Great Barrier Reef Marine Park (Version 2)}
\format{
A \href{https://cran.r-project.org/package=sf}{Simple Features data.frame}
with 6 rows and 9 variables:
\itemize{
\item OBJECTID:
\item NAME:
\item NRM_BODY:
\item COMMENTS:
\item IQATLAS_QL:
\item Area_KM2:
\item Shape_STAr:
\item Shape_STLe:
\item geometry:
}
}
\description{
This dataset was created to provide a maritime extension of the largely
terrestrial Natural Resource Management Regions. It will be used as a
planning tool when assessing water quality issues adjacent to the
Natural Resource Management Regions, within or adjacent to the Great
Barrier Reef Marine Park.
All data is based on the GDA2020 coordinate system.
}
\details{
A marine extension of the Natural Resource Management Regions, covering
an area from the Queensland mainland at low water, excluding all islands
and cays.

For full Metadata see the original
\href{https://tinyurl.com/yb3lgqf4}{eAtlas}
page info for the NRM areas flowing into the
Great Barrier Reef Marine Park.

Data kindly provided by:
Great Barrier Reef Marine Park Authority
Data Administrator
Spatial Data Centre
PO Box 1379
Townsville, QLD 4810
Australia
\email{gis@gbrmpa.gov.au}

The Great Barrier Reef Marine Park Authority requests attribution
in the following manner:
Copyright (c) Great Barrier Reef Marine Park Authority 2020.
This product is released under
the \href{http://creativecommons.org/licenses/by/4.0/legalcode}{Creative Commons Attribution 4.0 International Licence}.
Updated data available at \url{http://www.gbrmpa.gov.au/geoportal/}.
Under CC BY 4.0, users are free to share and adapt the
data under the following conditions:
\itemize{
\item They must give appropriate credit
\item Provide a link to the license
\item Indicate if any changes were made, and
\item They may not apply any additional restrictions that
legally restrict others from doing anything the license permits.
}
}
